RND-3308: fix reference to TI sources list
As part of building the Debian rootfs, we integrate the sources from
Texas Instruments.
Previously we could just include the latest version of the sources, but
these have been updates from bookworm to trixie, letting the build fail.

Reference a fixed version of that file to get the bookworm sources.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>